+++ ninja-core/src/site/markdown/developer/changelog.md
@@ -1,7 +1,7 @@
 Version X.X.X
 =============
 
- * 2014-08-09 BugFix #185. Rollback of #157. UnitOfWork heabior fixed.
+ * 2014-08-09 BugFix #185. Rollback of #157. UnitOfWork behavior fixed. (nobullet, cpisto, ra).
  * 2014-08-01 Implement proper 401 unauthorized support (gitblit)
  * 2014-08-01 Add an HTTP Basic Authentication filter and a UsernamePasswordValidator interface (gitblit)
  * 2014-08-01 Added type support for byte, char, short, enums, and arrays (gitblit)